requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
discord.py==1.7.3
futures==2.2.0
goslate==1.5.1
idna==3.3
multidict==5.2.0
PyDictionary==2.0.1
python-dotenv==0.19.1
requests==2.26.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
